+++ consensus/pbft/api/election_api.go
@@ -59,7 +59,7 @@ func (e *ElectionApi) Vote(connectionId string) error {
 	e.ElectionService.ResetLeftTime()
 
 	voteLeaderMessage := pbft.VoteMessage{}
-	grpcDeliverCommand, _ := CreateGrpcDeliverCommand("VoteLeaderProtocol", voteLeaderMessage)
+	grpcDeliverCommand, _ := common.CreateGrpcDeliverCommand("VoteLeaderProtocol", voteLeaderMessage)
 	grpcDeliverCommand.RecipientList = append(grpcDeliverCommand.RecipientList, connectionId)
 
 	iLogger.Infof(nil, "[PBFT] Vote to %s", connectionId)
@@ -75,7 +75,7 @@ func (e *ElectionApi) broadcastLeader(rep pbft.Representative) error {
 	updateLeaderMessage := pbft.UpdateLeaderMessage{
 		Representative: rep,
 	}
-	grpcDeliverCommand, err := CreateGrpcDeliverCommand("UpdateLeaderProtocol", updateLeaderMessage)
+	grpcDeliverCommand, err := common.CreateGrpcDeliverCommand("UpdateLeaderProtocol", updateLeaderMessage)
 	if err != nil {
 		iLogger.Errorf(nil, "[PBFT] Cannot create grpc command - Error: [%s]", err.Error())
 		return err
@@ -195,7 +195,7 @@ func (e *ElectionApi) RequestVote(peerIds []string) error {
 	requestVoteMessage := pbft.RequestVoteMessage{
 		Term: e.ElectionService.GetTerm(),
 	}
-	grpcDeliverCommand, _ := CreateGrpcDeliverCommand("RequestVoteProtocol", requestVoteMessage)
+	grpcDeliverCommand, _ := common.CreateGrpcDeliverCommand("RequestVoteProtocol", requestVoteMessage)
 
 	for _, connectionId := range peerIds {
 		grpcDeliverCommand.RecipientList = append(grpcDeliverCommand.RecipientList, connectionId)
@@ -233,20 +233,4 @@ func (e *ElectionApi) SetLeader(representativeId string) {
 	e.eventService.Publish("leader.updated", event.LeaderUpdated{
 		LeaderId: representativeId,
 	})
-}
-
-func CreateGrpcDeliverCommand(protocol string, body interface{}) (command.DeliverGrpc, error) {
-
-	data, err := common.Serialize(body)
-
-	if err != nil {
-		return command.DeliverGrpc{}, err
-	}
-
-	return command.DeliverGrpc{
-		MessageId:     xid.New().String(),
-		RecipientList: make([]string, 0),
-		Body:          data,
-		Protocol:      protocol,
-	}, err
 }
